Mais uma tentativa para passar no avaliador - mudança no arquivo 'CategoriesList.jsx'
